put FbGraph in it's own folder, move some files to `third-party`
