+++ shell/apps/utils/FBGraph/Field.js
@@ -1,5 +1,4 @@
-//import {FbGraph} from './FbGraph.js';
-import {FbGraph} from '../../../components/FbGraph.js';
+import {FbGraph} from '../../../components/FbGraph/FbGraph.js';
 import Firebase from '../../common/firebase-config.js';
 
 const {Field} = FbGraph(Firebase.db);
