+++ shell/apps/utils/FBGraph/elements/index.html
@@ -113,7 +113,7 @@
   </div>
 </div>
 
-<script src="../../../../components/renderjson.js"></script>
+<script src="../../../../components/third-party/renderjson.js"></script>
 <script>
   renderjson.set_icons(' ▶ ', ' ▼ ');
   renderjson.set_show_to_level(2);
