+++ shell/apps/vr/index.html
@@ -14,7 +14,6 @@
 
   <script src="https://aframe.io/releases/0.8.0/aframe.min.js"></script>
   <script src="https://unpkg.com/aframe-html-shader@0.2.0/dist/aframe-html-shader.min.js"></script>
-  <!-- <script type="module" src="../../components/a-frame/aframe-html.js"></script> -->
 
   <script type="module">
     // global configuration
@@ -25,7 +24,7 @@
     import '../../build/ArcsLib.js';
     // components for particle use
     import '../common/whitelisted.js';
-    import '../../components/a-frame/aframe-html.js';
+    import '../../components/third-party/aframe-html.js';
     // shell
     import './vr-app-shell.js';
   </script>
